+++ tests/app/ui/page/page-tests-common.ts
@@ -505,11 +505,6 @@ function _test_WhenInnerViewCallsCloseModal(closeModalGetter: (ShownModallyData)
     helper.navigate(masterPageFactory);
 
     TKUnit.waitUntilReady(() => modalClosedWithResult);
-
-    if (isIOS) {
-        // Remove this line when we have a good way to detect actual modal close on ios
-        TKUnit.waitUntilReady(() => !(<UIViewController>topmost().currentPage.viewController).presentedViewController);
-    }
 }
 
 export function test_WhenViewBaseCallsShowModal_WithArguments_ShouldOpenModal() {
@@ -572,11 +567,6 @@ export function test_WhenViewBaseCallsShowModal_WithArguments_ShouldOpenModal() 
     helper.navigate(masterPageFactory);
 
     TKUnit.waitUntilReady(() => modalClosed);
-
-    if (isIOS) {
-        // Remove this line when we have a good way to detect actual modal close on ios
-        TKUnit.waitUntilReady(() => !(<UIViewController>topmost().currentPage.viewController).presentedViewController);
-    }
 }
 
 export function test_WhenViewBaseCallsShowModal_WithShowModalOptionsArguments_ShouldOpenModal() {
@@ -794,11 +784,6 @@ export function test_WhenRootTabViewShownModallyItCanCloseModal() {
     helper.navigate(masterPageFactory);
 
     TKUnit.waitUntilReady(() => modalClosed);
-
-    if (isIOS) {
-        // Remove this line when we have a good way to detect actual modal close on ios
-        TKUnit.waitUntilReady(() => !(<UIViewController>topmost().currentPage.viewController).presentedViewController);
-    }
 }
 
 export function test_WhenPageIsNavigatedToItCanShowAnotherPageAsModal() {
@@ -885,11 +870,6 @@ export function test_WhenPageIsNavigatedToItCanShowAnotherPageAsModal() {
     TKUnit.assertEqual(modalUnloaded, 1, "modalUnloaded");
 
     masterPage.off(Page.navigatedToEvent, navigatedToEventHandler);
-
-    if (isIOS) {
-        // Remove this line when we have a good way to detect actual modal close on ios
-        TKUnit.waitUntilReady(() => !(<UIViewController>topmost().currentPage.viewController).presentedViewController);
-    }
 }
 
 export function test_WhenModalPageShownHostPageNavigationEventsShouldNotBeRaised() {
@@ -966,11 +946,6 @@ export function test_WhenModalPageShownHostPageNavigationEventsShouldNotBeRaised
     helper.navigate(masterPageFactory);
 
     TKUnit.waitUntilReady(() => ready);
-
-    if (isIOS) {
-        // Remove this line when we have a good way to detect actual modal close on ios
-        TKUnit.waitUntilReady(() => !(<UIViewController>topmost().currentPage.viewController).presentedViewController);
-    }
 
     // only raised by the initial navigation to the master page
     TKUnit.assertTrue(hostNavigatingToCount === 1);
@@ -1058,11 +1033,6 @@ export function test_WhenModalPageShownModalNavigationToEventsShouldBeRaised() {
 
     TKUnit.waitUntilReady(() => ready && !modalFrame.isLoaded);
 
-    if (isIOS) {
-        // Remove this line when we have a good way to detect actual modal close on ios
-        TKUnit.waitUntilReady(() => !(<UIViewController>topmost().currentPage.viewController).presentedViewController);
-    }
-
     // only raised by the initial show modal navigation
     TKUnit.assertTrue(modalNavigatingToCount === 1);
     TKUnit.assertTrue(modalNavigatedToCount === 1);
@@ -1138,12 +1108,6 @@ export function test_WhenModalFrameShownModalEventsRaisedOnRootModalFrame() {
     helper.navigate(masterPageFactory);
 
     TKUnit.waitUntilReady(() => ready && !modalFrame.isLoaded);
-
-    if (isIOS) {
-        // Remove this line when we have a good way to detect actual modal close on ios
-        TKUnit.waitUntilReady(() => !(<UIViewController>topmost().currentPage.viewController).presentedViewController);
-    }
-
     TKUnit.assertTrue(showingModallyCount === 1);
     TKUnit.assertTrue(shownModallyCount === 1);
 }
@@ -1206,12 +1170,6 @@ export function test_WhenModalPageShownShowModalEventsRaisedOnRootModalPage() {
     helper.navigate(masterPageFactory);
 
     TKUnit.waitUntilReady(() => ready);
-
-    if (isIOS) {
-        // Remove this line when we have a good way to detect actual modal close on ios
-        TKUnit.waitUntilReady(() => !(<UIViewController>topmost().currentPage.viewController).presentedViewController);
-    }
-
     TKUnit.assertTrue(showingModallyCount === 1);
     TKUnit.assertTrue(shownModallyCount === 1);
 }
@@ -1279,12 +1237,6 @@ export function test_WhenModalPageShownShowModalEventsRaisedOnRootModalTabView()
     TKUnit.assertEqual(_stack().length, 2, "Host and modal tab frame should be instantiated at this point!");
 
     TKUnit.waitUntilReady(() => ready);
-
-    if (isIOS) {
-        // Remove this line when we have a good way to detect actual modal close on ios
-        TKUnit.waitUntilReady(() => !(<UIViewController>topmost().currentPage.viewController).presentedViewController);
-    }
-
     TKUnit.assertEqual(_stack().length, 1, "Single host frame should be instantiated at this point!");
 
     TKUnit.assertTrue(showingModallyCount === 1);